Move Directive to /checkers/todo_directives/TodoDirective
